refactor(team): companion object Mapper 클래스로 추출
