+++ Backend/src/main/java/com/luckyseven/backend/domain/team/dto/TeamCreateResponse.kt
@@ -7,18 +7,4 @@ data class TeamCreateResponse(
     val name: String?,
     val teamCode: String?,
     val leaderId: Long?
-) {
-    companion object {
-        fun toTeamCreateResponse(team: Team?): TeamCreateResponse? {
-            return team?.let {
-                TeamCreateResponse(
-                    id = it.id,
-                    name = it.name,
-                    teamCode = it.teamCode,
-                    leaderId = it.leader?.id
-                )
-            }
-        }
-
-    }
-}
+)
