+++ Backend/src/main/java/com/luckyseven/backend/domain/team/dto/TeamJoinResponse.kt
@@ -7,17 +7,4 @@ data class TeamJoinResponse(
     val teamName: String?,
     val teamCode: String?,
     val leaderId: Long?
-) {
-    companion object {
-        fun toTeamJoinResponse(team: Team?): TeamJoinResponse? {
-            return team?.let {
-                TeamJoinResponse(
-                    id = it.id,
-                    teamName = it.name,
-                    teamCode = it.teamCode,
-                    leaderId = it.leader?.id
-                )
-            }
-        }
-    }
-}
+)
