+++ Backend/src/main/java/com/luckyseven/backend/domain/team/dto/TeamListResponse.kt
@@ -6,14 +6,4 @@ data class TeamListResponse(
     val id: Long?,
     val name: String?,
     val teamCode: String?
-) {
-    companion object{
-        fun toTeamListResponse(team: Team?): TeamListResponse?{
-            return TeamListResponse(
-                id = team?.id,
-                name = team?.name,
-                teamCode = team?.teamCode
-            )
-        }
-    }
-}
+)
