+++ Backend/src/main/java/com/luckyseven/backend/domain/team/dto/TeamMemberDto.kt
@@ -25,23 +25,6 @@ data class TeamMemberDto(
     val role: String?
 ) {
     companion object {
-        fun toTeamMemberDto(teamMember: TeamMember?): TeamMemberDto? {
-            if (teamMember == null) return null
 
-            val team = teamMember.team
-            val member = teamMember.member
-            val leader = team?.leader
-            val role = if (leader?.id == member?.id) "Leader" else "Member"
-
-            return TeamMemberDto(
-                id = teamMember.id,
-                teamId = team?.id,
-                teamName = team?.name,
-                memberId = member?.id,
-                memberNickName = member?.nickname,
-                memberEmail = member?.email,
-                role = role
-            )
-        }
     }
 }
